requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0
scikit-image>=0.18.0
imageio>=2.9.0
torch>=2.0.0
torchvision>=0.15.0
PyYAML>=5.4
matplotlib>=3.3.0
